Fix cover download.
Ensure that the cache directory for covers
exists.

Also, properly reload the artist image in artist
list when a cover is associated to an artist.

Rewrite the text position on cover with a direct
calculation of the Y position, the anchors solution
with alternating undefined values seems not to work
properly, when a cover is associated to an artist
for instance.