[utils.cpp] Avoid creating QSettings too early (#65)
If the QSettings are initialised statically they may fail to pick up the
organization and application names set in main. The result is that the
settings are written to and read from both
~~"~/.config/Unknown\Organization.conf" and~~
"~/.config/sailfishos-applications/flowplayer.conf".

This can prevent the configured Music directories being read from
successfully.

P.S.: We can still move to a singleton object later if needed.