Update README.md
This is a base for a new README with all the essential features I could think of. I don't have a working phone right now, so the listing is not perfect.

I have marked all the bits I was uncertain about with a question mark, so please feel free to correct, delete and add anything you see fit.